working version before making db dynamic

+++ frontend2/src/app/model_input/input-field/input-field.component.ts
@@ -32,7 +32,8 @@ export class InputFieldComponent implements OnInit {
   };
 
   valChanged(newVal: string){
-    const newValNumber = parseInt(newVal)
+  
+    const newValNumber = parseFloat(newVal)
     this.onChange.emit(newValNumber);
     console.log("VALCHANGEDDDDD")
     console.log(newValNumber)
